+++ CHANGELOG.md
@@ -6,11 +6,14 @@
 * `values`, `keys`, `set`, `remove` to use objects as observable collections
 * Introduced `onBecomeObserved` and `onBecomeUnobserved`
 * MobX now supports development only checks and exceptions, resulting in smaller and faster production builds. The setup requirements are identical to react
+* added `requiresReaction` option to computed
 * Introduced `mobx.configure({ computedRequiresReaction: true })`
 * Introduced `mobx.configure({ disableErrorBoundaries })`, for easier debugging of exceptoins. By [NaridaL](https://github.com/NaridaL) through [#1262](https://github.com/mobxjs/mobx/pull/1262)
+* Introduced `@computed(options)`
 
 ## Breaking changes
 
+* dropped `@computed.equals`, instead, you can now use `@computed({ equals: ... })`
 * `useStrict(boolean)` was dropped, use `configure({enforceActions: boolean})` instead
 * `isolateGlobalState` was dropped, use `configure({ isolateGlobalState: true})` instead
 * `expr` is moved to mobx-utils. Remember, `expr(fn)` is just sugar for `computed(fn).get()`